+++ src/EventStore.Core.Tests/Helpers/IODispatcherTests/ReadEventsTests/read_stream_events_forward_with_successful_read.cs
@@ -6,7 +6,7 @@ using NUnit.Framework;
 
 namespace EventStore.Core.Tests.Helpers.IODispatcherTests.ReadEventsTests {
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	public class async_read_stream_events_forward_with_successful_read<TLogFormat, TStreamId> : with_read_io_dispatcher<TLogFormat, TStreamId> {
 		private ClientMessage.ReadStreamEventsForwardCompleted _result;
 		private bool _hasTimedOut;
@@ -50,7 +50,7 @@ namespace EventStore.Core.Tests.Helpers.IODispatcherTests.ReadEventsTests {
 	}
 
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	public class read_stream_events_forward_with_successful_read<TLogFormat, TStreamId> : with_read_io_dispatcher<TLogFormat, TStreamId> {
 		private ClientMessage.ReadStreamEventsForwardCompleted _result;
 		private bool _hasTimedOut;
