+++ src/EventStore.Core.Tests/Http/PersistentSubscription/creating.cs
@@ -11,7 +11,7 @@ using EventStore.Transport.Http;
 namespace EventStore.Core.Tests.Http.PersistentSubscription {
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_a_subscription<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		private HttpResponseMessage _response;
 
@@ -44,7 +44,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_a_subscription_with_query_params<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		private HttpResponseMessage _response;
 
@@ -77,7 +77,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_a_subscription_without_permissions<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		private HttpResponseMessage _response;
 
@@ -105,7 +105,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_a_duplicate_subscription<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		private HttpResponseMessage _response;
 
@@ -138,7 +138,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_a_subscription_with_bad_config<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		protected List<object> Events;
 		protected string SubscriptionPath;
@@ -185,7 +185,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_persistent_subscription_with_message_timeout_0<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		protected string SubscriptionPath;
 		protected string GroupName;
@@ -227,7 +227,7 @@ namespace EventStore.Core.Tests.Http.PersistentSubscription {
 
 	[Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	class when_creating_persistent_subscription_without_message_timeout<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 		protected string SubscriptionPath;
 		protected string GroupName;
